requirements.txt
ansible==2.0.1.0
bernhard==0.2.4
influxdb==2.12.0
python-decouple==3.0
